+++ lib/nxt_support/types.rb
@@ -1,0 +1,2 @@
+require "nxt_support/types/indifferent_json_type"
+require "nxt_support/types/encrypted_json_paths_type"